backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Ali Rachidi", email = "rachidiali10@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[tool.poetry.dependencies]
django = "^5.1.4"
djangorestframework = "^3.15.2"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.10"
